requirements.txt
Django
django-colorfield
Pillow
xhtml2pdf